requirements.txt

Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
celery==5.3.4
redis==4.5.4
django-cors-headers==4.2.0